% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetRainbowBar.R
\name{SetRainbowBar}
\alias{SetRainbowBar}
\title{Set RainbowBar Starting Variables}
\usage{
SetRainbowBar(min, max)
}
\arguments{
\item{min}{Starting Tick}

\item{max}{Ending Tick}
}
\description{
Set RainbowBar Starting Variables
}
